Parallelizing get_all_forecasts speeds up download
This uses requests_futures to parallalelize
get_avalanche_warnings_as_json(). It still just downloads one region
at a time, but now it also only downloads 31 days per API call. This is
because the API gets really slow when a whole season is requested
at once.

The error handling is also modified to better accomodate the async model
of requests_futures. Since it is now handled iteratively, the parameter
name recursive_count is not technically correct, but I did not want to
change the function API.