Client for submitting observations to Regobs
Currently just a proof-of-concept. The only observations that are
supported are Danger Signs and Notes. However, it should be easy
to add more.

It's currently using Regobs v4, since the OpenID Connect solution
for v5 requires the app to be registered. Hopefully that can be
solved in the future. In the meantime, we have to use the old
username/password combination for the accounts. It's still possible
to register new "old" accounts, so it shouldn't pose any practical
difficulties for any one.

See row 199 for an example of how to use the client.